+++ src/pages/apps/index.tsx
@@ -8,6 +8,13 @@ import SafeAppsSDKLink from '@/components/safe-apps/SafeAppsSDKLink'
 import SafeAppsHeader from '@/components/safe-apps/SafeAppsHeader'
 import SafeAppList from '@/components/safe-apps/SafeAppList'
 import { AppRoutes } from '@/config/routes'
+import type { GetServerSideProps } from 'next'
+import { checkForForbdidenRegions } from '@/utils/geo'
+
+export const getServerSideProps: GetServerSideProps<{}> = async ({ req, res }) => {
+  checkForForbdidenRegions(req, res)
+  return { props: {} }
+}
 
 const SafeApps: NextPage = () => {
   const router = useRouter()
