+++ src/pages/transactions/tx.tsx
@@ -1,8 +1,10 @@
-import type { NextPage } from 'next'
+import type { NextPage, GetServerSideProps } from 'next'
+
 import Head from 'next/head'
 
 import SingleTx from '@/components/transactions/SingleTx'
 import Typography from '@mui/material/Typography'
+import { checkForForbdidenRegions } from '@/utils/geo'
 
 const SingleTransaction: NextPage = () => {
   return (
@@ -23,3 +25,8 @@ const SingleTransaction: NextPage = () => {
 }
 
 export default SingleTransaction
+
+export const getServerSideProps: GetServerSideProps<{}> = async ({ req, res }) => {
+  checkForForbdidenRegions(req, res)
+  return { props: {} }
+}
