+++ src/pages/welcome.tsx
@@ -1,6 +1,9 @@
-import type { NextPage } from 'next'
+import type { NextPage, GetServerSideProps } from 'next'
+
 import Head from 'next/head'
 import NewSafe from '@/components/welcome/NewSafe'
+
+import { checkForForbdidenRegions } from '@/utils/geo'
 
 const Welcome: NextPage = () => {
   return (
@@ -15,3 +18,8 @@ const Welcome: NextPage = () => {
 }
 
 export default Welcome
+
+export const getServerSideProps: GetServerSideProps<{}> = async ({ req, res }) => {
+  checkForForbdidenRegions(req, res)
+  return { props: {} }
+}
